Combine related file operations to a dedicated app
File scanning and private files features were closely related.
This way the interplay between the features should be more
understandable by proximity.
