+++ file_operations/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class FilescanConfig(AppConfig):
+class FileOperationsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "filescan"
+    name = "file_operations"
