--- filescan/tests/example_response.json
@@ -1,19 +1,0 @@
-{
-  "success": true,
-  "data": {
-    "result": [
-      {
-        "name": "1Mfile01.rnd",
-        "is_infected": false,
-        "viruses": []
-      },
-      {
-        "name": "eicar_com.zip",
-        "is_infected": true,
-        "viruses": [
-          "Win.Test.EICAR_HDB-1"
-        ]
-      }
-    ]
-  }
-}
