+++ leasing/models/debt_collection.py
@@ -7,10 +7,10 @@ from django.utils.translation import pgettext_lazy
 from docxtpl import DocxTemplate
 
 from field_permissions.registry import field_permissions
-from filescan.mixins import FileScanMixin
+from file_operations.private_files import PrivateFileField
+from file_operations.viewsets.mixins import FileScanMixin
 from leasing.models.mixins import TimeStampedSafeDeleteModel
 from users.models import User
-from utils.models.fields import PrivateFileField
 
 
 def get_collection_letter_file_upload_to(instance, filename):
