+++ leasing/models/infill_development_compensation.py
@@ -5,12 +5,12 @@ from django.utils.translation import pgettext_lazy
 from enumfields import EnumField
 
 from field_permissions.registry import field_permissions
-from filescan.mixins import FileScanMixin
+from file_operations.private_files import PrivateFileField
+from file_operations.viewsets.mixins import FileScanMixin
 from leasing.enums import InfillDevelopmentCompensationState
 from leasing.models.decision import DecisionMaker
 from leasing.models.lease import IntendedUse
 from users.models import User
-from utils.models.fields import PrivateFileField
 
 from .mixins import TimeStampedSafeDeleteModel
